refactor: move shard_layout_from_prev_block to EMA (#7673)
Part of https://github.com/near/nearcore/issues/6910
