+++ tools/state-viewer/src/apply_chunk.rs
@@ -408,6 +408,7 @@ mod test {
     use near_chain_configs::Genesis;
     use near_client::test_utils::TestEnv;
     use near_crypto::{InMemorySigner, KeyType};
+    use near_epoch_manager::EpochManagerAdapter;
     use near_network::types::NetworkClientResponses;
     use near_primitives::hash::CryptoHash;
     use near_primitives::runtime::config_store::RuntimeConfigStore;
